requirements.txt
absl-py==2.1.0
anyio==4.3.0
appdirs==1.4.4
ascendctools==0.1.0
astor==0.8.1
astroid==3.1.0
asttokens==2.4.1
attrs==23.2.0
audioread==3.0.1
auto-tune==0.1.0
autograd==1.4
Babel==2.14.0
bce-python-sdk==0.9.7
blinker==1.8.1
certifi==2019.11.28
cffi==1.16.0
cfgv==3.4.0
chardet==3.0.4
clang-format==13.0.0
click==8.1.7
cloudpickle==3.0.0
comm==0.2.2
contourpy==1.2.1
coverage==5.5
cryptography==42.0.5
cycler==0.12.1
Cython==3.0.10
dataflow==0.0.1
dbus-python==1.2.16
debugpy==1.8.1
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.8
distlib==0.3.8
distro==1.4.0
distro-info==0.23+ubuntu1.1
docker-pycreds==0.4.0
easydict==1.13
exceptiongroup==1.2.1
executing==2.0.1
faiss-cpu==1.8.0.post1
Farama-Notifications==0.0.4
filelock==3.14.0
Flask==3.0.3
flask-babel==4.0.0
fonttools==4.51.0
future==1.0.0
gast==0.3.3
gitdb==4.0.11
GitPython==3.1.43
gym==0.26.2
gym-notices==0.0.8
gymnasium==1.1.1
h11==0.14.0
hccl==0.1.0
hccl-parser==0.1
httpcore==1.0.5
httpx==0.27.0
hypothesis==6.100.2
identify==2.5.36
idna==2.8
importlib_metadata==7.1.0
importlib_resources==6.4.0
iniconfig==2.0.0
ipykernel==6.29.4
ipython==8.18.1
isort==5.13.2
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.0
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
librosa==0.8.1
llvmlite==0.42.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mccabe==0.7.0
mock==5.1.0
mpi4py==4.0.2
mpmath==1.3.0
msadvisor==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
nodeenv==1.8.0
numba==0.59.1
numpy==1.24.4
op-compile-tool==0.1.0
op-gen==0.1
op-test-frame==0.1
opc-tool==0.1.0
opencv-python==3.4.18.65
opt-einsum==3.3.0
packaging==24.0
paddle-custom-npu==0.0.0
paddle2onnx==1.0.5
paddlepaddle==3.0.0
pandas==2.2.2
parameterized==0.9.0
parso==0.8.4
pathlib2==2.3.7.post1
pettingzoo==1.24.3
pexpect==4.9.0
pillow==10.3.0
pip==24.0
platformdirs==4.2.1
pluggy==1.5.0
pooch==1.8.1
pre-commit==2.17.0
prettytable==3.10.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
pycrypto==2.6.1
pycryptodome==3.20.0
pygame==2.5.2
PyGithub==2.3.0
Pygments==2.17.2
PyGObject==3.36.0
PyJWT==2.8.0
pylint==3.1.0
PyNaCl==1.5.0
pyparsing==3.1.2
pytest==8.2.0
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.2
qtconsole==5.5.1
QtPy==2.4.1
rarfile==4.2
requests==2.22.0
requests-unixsocket==0.2.0
resampy==0.4.3
schedule-search==0.0.1
scikit-learn==1.4.2
scipy==1.13.0
sentry-sdk==2.0.1
setproctitle==1.3.3
setuptools==70.0.0
six==1.14.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
ssh-import-id==5.10
stack-data==0.6.3
sympy==1.12
te==0.4.0
threadpoolctl==3.5.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.4
tornado==6.4
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2024.1
ubelt==1.3.3
ujson==5.10.0
unattended-upgrades==0.1
urllib3==2.2.1
virtualenv==20.26.1
visualdl==2.5.3
wandb==0.16.6
wcwidth==0.2.13
Werkzeug==3.0.2
wheel==0.43.0
wrapt==1.16.0
xdoctest==1.1.1
XlsxWriter==3.0.9
xuance==1.2.4
zipp==3.18.1
